+++ src/lib/pages/workflow-query.svelte
@@ -4,9 +4,9 @@
   import { workflowRun } from '$lib/stores/workflow-run';
 
   import CodeBlock from '$lib/holocene/code-block.svelte';
-  import Option from '$lib/components/select/option.svelte';
+  import Select from '$lib/holocene/select/simple-select.svelte';
+  import Option from '$lib/holocene/select/simple-option.svelte';
   import EmptyState from '$lib/holocene/empty-state.svelte';
-  import Select from '$lib/components/select/select.svelte';
   import Button from '$holocene/button.svelte';
   import PageTitle from '$lib/holocene/page-title.svelte';
   import Loading from '$lib/holocene/loading.svelte';
